+++ taker/src/routes.rs
@@ -131,6 +131,7 @@ impl Heartbeat {
 pub struct CfdOrderRequest {
     pub order_id: OrderId,
     pub quantity: Usd,
+    pub leverage: Leverage,
 }
 
 #[rocket::post("/cfd/order", data = "<cfd_order_request>")]
@@ -139,8 +140,18 @@ pub async fn post_order_request(
     taker: &State<Taker>,
     _auth: Authenticated,
 ) -> Result<(), HttpApiProblem> {
+    // TODO: remove me. This is a temporary fix to ensure no one tries to use a different leverage
+    // for now
+    if cfd_order_request.leverage != Leverage::TWO {
+        return Err(HttpApiProblem::new(StatusCode::BAD_REQUEST)
+            .title("Invalid leverage selected. For time being, please go with leverage 2."));
+    }
     taker
-        .take_offer(cfd_order_request.order_id, cfd_order_request.quantity)
+        .take_offer(
+            cfd_order_request.order_id,
+            cfd_order_request.quantity,
+            cfd_order_request.leverage,
+        )
         .await
         .map_err(|e| {
             HttpApiProblem::new(StatusCode::INTERNAL_SERVER_ERROR)